requirements.txt
asgiref==3.5.0
boto3==1.21.25
botocore==1.24.25
Django==4.0.3
django-cors-headers==3.11.0
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
jmespath==1.0.0
Pillow==9.0.1
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2022.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.0.0
